comment cleanup; move descaletensors to utils
