+++ flash_attn/cute/flash_bwd.py
@@ -30,8 +30,6 @@ from flash_attn.cute.utils import AuxData
 
 
 class FlashAttentionBackwardSm80:
-    # NamedTuple requires fields without defaults first; every use is by name, so the order
-    # carries no meaning beyond that.
     class Args(NamedTuple):
         mQ: cute.Tensor
         mK: cute.Tensor
@@ -400,7 +398,7 @@ class FlashAttentionBackwardSm80:
     @cute.jit
     def __call__(
         self,
-        args,  # Args, or any namedtuple whose extra fields are all None
+        args,
         # Always keep stream as the last parameter (EnvStream: obtained implicitly via TVM FFI).
         stream: cuda.CUstream = None,
     ):
