+++ flash_attn/cute/flash_bwd_sm90.py
@@ -45,8 +45,6 @@ from flash_attn.cute.block_sparse_utils import (
 
 
 class FlashAttentionBackwardSm90:
-    # NamedTuple requires fields without defaults first; every use is by name, so the order
-    # carries no meaning beyond that.
     class Args(NamedTuple):
         mQ: cute.Tensor
         mK: cute.Tensor
@@ -370,7 +368,7 @@ class FlashAttentionBackwardSm90:
     @cute.jit
     def __call__(
         self,
-        args,  # Args, or any namedtuple whose extra fields are all None
+        args,
         # Always keep stream as the last parameter (EnvStream: obtained implicitly via TVM FFI).
         stream: cuda.CUstream = None,
     ):
