+++ flash_attn/cute/flash_fwd.py
@@ -312,7 +312,7 @@ class FlashAttentionForwardBase:
     @cute.jit
     def __call__(
         self,
-        args,  # the subclass's Args, or any namedtuple whose extra fields are all None
+        args,
         # Always keep stream as the last parameter (EnvStream: obtained implicitly via TVM FFI).
         stream: cuda.CUstream = None,
     ):
@@ -573,8 +573,6 @@ class FlashAttentionForwardBase:
 
 
 class FlashAttentionForwardSm80(FlashAttentionForwardBase):
-    # NamedTuple requires fields without defaults first; every use is by name, so the order
-    # carries no meaning beyond that.
     class Args(NamedTuple):
         mQ: cute.Tensor
         mK: cute.Tensor
@@ -640,7 +638,7 @@ class FlashAttentionForwardSm80(FlashAttentionForwardBase):
     @cute.jit
     def __call__(
         self,
-        args,  # Args, or any namedtuple whose extra fields are all None
+        args,
         # Always keep stream as the last parameter (EnvStream: obtained implicitly via TVM FFI).
         stream: cuda.CUstream = None,
     ):
