+++ flash_attn/cute/flash_fwd_sm90.py
@@ -51,8 +51,6 @@ from flash_attn.cute.utils import AuxData
 
 
 class FlashAttentionForwardSm90(FlashAttentionForwardBase):
-    # NamedTuple requires fields without defaults first; every use is by name, so the order
-    # carries no meaning beyond that.
     class Args(NamedTuple):
         mQ: cute.Tensor  # (b, s_q, h, d) or (total_q, h, d) if there is cu_seqlens_q
         mK: cute.Tensor  # (b_k, s_k, h_k, d), (total_k, h_k, d) with cu_seqlens_k, or (num_pages, page_size, h_k, d) with page_table
@@ -180,7 +178,7 @@ class FlashAttentionForwardSm90(FlashAttentionForwardBase):
     @cute.jit
     def __call__(
         self,
-        args,  # Args, or any namedtuple whose extra fields are all None
+        args,
         # Always keep stream as the last parameter (EnvStream: obtained implicitly via TVM FFI).
         stream: cuda.CUstream = None,
     ):
