+++ flash_attn/cute/interface.py
@@ -39,7 +39,7 @@ from flash_attn.cute.cute_dsl_utils import (
 )
 from flash_attn.cute.flash_fwd import FlashAttentionForwardSm80
 from flash_attn.cute.flash_fwd_sm90 import FlashAttentionForwardSm90
-from flash_attn.cute.flash_fwd_sm100 import FlashAttentionForwardSm100, DescaleTensors
+from flash_attn.cute.flash_fwd_sm100 import FlashAttentionForwardSm100
 from flash_attn.cute.flash_fwd_sm120 import FlashAttentionForwardSm120
 from flash_attn.cute.flash_bwd_preprocess import FlashAttentionBackwardPreprocess
 from flash_attn.cute.flash_bwd import FlashAttentionBackwardSm80
@@ -62,7 +62,7 @@ from flash_attn.cute.flash_bwd_mla_dk_sm100 import dKGemmKernel
 from flash_attn.cute.sm100_hd256_2cta_fmha_forward import BlackwellFusedMultiHeadAttentionForward
 from flash_attn.cute.sm100_hd256_2cta_fmha_backward import BlackwellFusedMultiHeadAttentionBackward
 
-from flash_attn.cute.utils import AuxData
+from flash_attn.cute.utils import AuxData, DescaleTensors
 from flash_attn.cute.block_sparsity import (
     BlockSparseTensorsTorch,
     block_sparse_bwd_supports_2cta,
@@ -2848,8 +2848,6 @@ def _flash_attn_bwd_sparse_mla(
         seqused_q,
         seqused_k,
     ) -> FlashAttentionSparseMLABackwardSm100.Args:
-        """Build the kernel arguments; used for both compilation and invocation so that the
-        two can never disagree on which arguments are present."""
         return FlashAttentionSparseMLABackwardSm100.Args(
             mdO=do,
             mV=v,
